refactor: split ups_rs_client.py into focused mixin modules
Extract WebSocket, async operations, event management, response
parsing, and workitem utility code into dedicated mixin modules to
bring all files under 600 lines. UPSRSClient now inherits from
WebSocketMixin, AsyncOperationsMixin, EventManagementMixin,
ResponseParsingMixin, and WorkitemUtilsMixin. Public API is
unchanged and all 148 tests continue to pass.
